#50 Added code generator for `CompanyIdentifier`
