+++ Tests/BluetoothTests/BluetoothUUIDTests.swift
@@ -27,180 +27,6 @@ final class BluetoothUUIDTests: XCTestCase {
     
     lazy var randomUUIDs = (1 ... 100000).map { _ in UUID() }
     
-    #if os(macOS) && swift(>=3.2)
-    func testGenerateDefinedUUIDExtension() {
-        
-        let uuids = definedUUIDs.sorted(by: { $0.0.key < $0.1.key })
-        
-        var generatedCode = ""
-        
-        var memberNameCache = [UInt16: String]()
-        
-        func 🖨(_ text: String) {
-            
-            generatedCode += text + "\n"
-        }
-        
-        // https://gist.github.com/AmitaiB/bbfcba3a21411ee6d3f972320bcd1ecd
-        func camelCase(_ string: String) -> String {
-            return string.components(separatedBy: CharacterSet.alphanumerics.inverted)
-                .filter { !$0.isEmpty }
-                .map { $0.capitalized }
-                .joined()
-        }
-        
-        func llamaCase(_ string: String) -> String {
-            var result = camelCase(string)
-            if let firstLetterCharacter = result.first {
-                result = String(result.dropFirst())
-                let firstLetter = String(firstLetterCharacter)
-                result = firstLetter.lowercased() + result
-            }
-            return result
-        }
-        
-        func uppercaseFirstLetter(_ string: String) -> String {
-            var result = string
-            if let firstLetterCharacter = result.first {
-                result = String(result.dropFirst())
-                let firstLetter = String(firstLetterCharacter)
-                result = firstLetter.uppercased() + result
-            }
-            return result
-        }
-        
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateStyle = .short
-        dateFormatter.timeStyle = .none
-        
-        let fileDate = dateFormatter.string(from: Date())
-        
-        🖨("//")
-        🖨("//  DefinedUUIDExtension.swift")
-        🖨("//  Bluetooth")
-        🖨("//")
-        🖨("//  Generated by Alsey Coleman Miller on \(fileDate).")
-        🖨("//")
-        🖨("")
-        🖨("public extension BluetoothUUID {")
-        🖨("")
-        
-        for (uuidValue, name) in uuids {
-            
-            let uuid = BluetoothUUID.bit16(uuidValue)
-            
-            let sanitizedName = name
-                .replacingOccurrences(of: "3D ", with: "uuid3D")
-                .replacingOccurrences(of: "360", with: "uuid360")
-                .replacingOccurrences(of: "3M", with: "uuid3M")
-                .replacingOccurrences(of: "Amazon.com Services, Inc.", with: "Amazon")
-                .replacingOccurrences(of: "The ", with: "")
-                .replacingOccurrences(of: "A/V", with: "av")
-                .replacingOccurrences(of: " AG & Co. KG", with: "")
-                .replacingOccurrences(of: " AG & Co.", with: "")
-                .replacingOccurrences(of: "Co.,", with: "")
-                .replacingOccurrences(of: "Co.,Ltd", with: "")
-                .replacingOccurrences(of: " AG", with: "")
-                .replacingOccurrences(of: " Sp. z o.o.", with: "")
-                .replacingOccurrences(of: " ASA", with: "")
-                .replacingOccurrences(of: " d.o.o.", with: "")
-                .replacingOccurrences(of: " SA", with: "")
-                .replacingOccurrences(of: " gmbh", with: "")
-                .replacingOccurrences(of: " GmbH", with: "")
-                .replacingOccurrences(of: " B.V.", with: "")
-                .replacingOccurrences(of: ",Inc.", with: "")
-                .replacingOccurrences(of: " Inc", with: "")
-                .replacingOccurrences(of: " INC", with: "")
-                .replacingOccurrences(of: " LLC", with: "")
-                .replacingOccurrences(of: " LTD", with: "")
-                .replacingOccurrences(of: " Ltd", with: "")
-                .replacingOccurrences(of: " A/S", with: "")
-            
-            let llamaCaseName = llamaCase(sanitizedName)
-            
-            var memberName = llamaCaseName
-            
-            // prevent duplicate entries
-            var duplicateNumber = 1
-            while memberNameCache.values.contains(memberName) {
-                
-                duplicateNumber += 1
-                memberName = llamaCaseName + "\(duplicateNumber)"
-            }
-            
-            let comment = name + " " + "(`0x\(uuid.rawValue)`)"
-            
-            🖨("    /// " + comment)
-            🖨("    static var " + memberName + ": BluetoothUUID {")
-            🖨("        return .bit16(0x" + uuid.rawValue + ")")
-            🖨("    }")
-            🖨("")
-            
-            memberNameCache[uuidValue] = memberName
-        }
-        
-        🖨("}")
-        
-        var filename = NSTemporaryDirectory() + "DefinedUUIDExtension.swift"
-        XCTAssertNoThrow(try generatedCode.write(toFile: filename, atomically: true, encoding: .utf8))
-        
-        print("Generated Swift code \(filename)")
-        
-        // generate unit test for extensions
-        generatedCode = ""
-        
-        🖨("//")
-        🖨("//  DefinedUUIDTests.swift")
-        🖨("//  Bluetooth")
-        🖨("//")
-        🖨("//  Generated by Alsey Coleman Miller on \(fileDate).")
-        🖨("//")
-        🖨("")
-        🖨("import XCTest")
-        🖨("import Foundation")
-        🖨("@testable import Bluetooth")
-        🖨("")
-        🖨("final class DefinedUUIDTests: XCTestCase {")
-        🖨("")
-        🖨("    static let allTests = [")
-        🖨("        (\"testDefinedUUID\", testDefinedUUID)")
-        🖨("    ]")
-        🖨("")
-        🖨("    func testDefinedUUID() {")
-        🖨("")
-        
-        
-        // generate test methods
-        
-        for (uuidValue, name) in uuids {
-            
-            let uuid = BluetoothUUID.bit16(uuidValue)
-            
-            guard let memberName = memberNameCache[uuidValue]
-                else { XCTFail("No extension generated for \(uuid)"); return }
-            
-            
-            🖨("        /// \(name)")
-            🖨("        XCTAssertEqual(BluetoothUUID.\(memberName).rawValue, \"\(uuid.rawValue)\")")
-            🖨("        XCTAssertEqual(BluetoothUUID.\(memberName), .bit16(0x\(uuid.rawValue)))")
-            🖨("        XCTAssertEqual(BluetoothUUID.\(memberName), .bit16(\(uuidValue)))")
-            🖨("        XCTAssertEqual(BluetoothUUID.\(memberName).name, \"\(name)\")")
-            🖨("        XCTAssertNotEqual(BluetoothUUID.\(memberName), .bit32(\(uuidValue)))")
-            🖨("        XCTAssertNotEqual(BluetoothUUID.\(memberName), .bit32(0x\(uuid.rawValue)))")
-            🖨("")
-        }
-        
-        🖨("    }")
-        🖨("")
-        🖨("}")
-        
-        filename = NSTemporaryDirectory() + "DefinedUUIDTests.swift"
-        XCTAssertNoThrow(try generatedCode.write(toFile: filename, atomically: true, encoding: .utf8))
-        
-        print("Generated Swift code \(filename)")
-    }
-    #endif
-    
     func testMalformed() {
         
         let malformed = [
